+++ CHANGES.rst
@@ -7,6 +7,7 @@ Changes
 - Based filters on file names which simplifies config a lot
 - Breaking change in YAML file config format
 - Breaking change in git filter config (one config to rule all secrets)
+- Changed YAML parser to ruamel.yaml
 
 0.0 (unreleased)
 ----------------